task1/requirements.txt
pytest>=7.0
PyYAML>=6.0
requests>=2.31
